requirements.txt
numpy>=1.21.0
pandas>=1.3.0
duckdb>=0.8.0
faiss-cpu>=1.7.0
annoy>=1.17.0
matplotlib>=3.4.0
seaborn>=0.11.0
psutil>=5.8.0
tqdm>=4.62.0
pyyaml>=5.4.0
scikit-learn>=1.0.0
pyarrow>=6.0.0
scipy>=1.7.0
ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
reportlab>=3.6.0

# Agent frameworks (optional - agents fallback gracefully if not installed)
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
pyautogen>=0.2.0

